"devel" branch started with the major code refactoring
 * split source files into smaller files.
   - copyright.py still needs to be worked on.
 * -x[01234] supported
 * -o dropped because we do not do it.
 * -y added to confirm before deleting the source tree.
 * multi-arch and hardening turned on
 * configurable debug with DEBUG=iof

Now, the debmake source develpment is hosted in the "devel" branch.
"debmake -d" is used to create the source tarball of the Debian
non-native package without debian/* files and the source tree under the
debmake-<version>/ directory.

This development workflow works almost like the native package one
except for the package version string which always has the Debian
revision -1 attached to it.

"upstream" and "master" brances are kept as typical contents of
git-buildpackage of the non-native Debian package.  (the same for
"pristine-tar")  They will be updated by the generated and uploaded
Debian package.